refactor: вынести чек ширины экрана в отдельный хук
